otra vez error en crear factura

+++ G5_Avance2_SC-504/src/main/java/com/frenchies/g5_avance2_sc504/service/FacturaService.java
@@ -1,4 +1,3 @@
-// src/main/java/com/frenchies/g5_avance2_sc504/service/FacturaService.java
 package com.frenchies.g5_avance2_sc504.service;
 
 import java.util.HashMap;
@@ -41,33 +40,24 @@ public class FacturaService {
         facturaRepo.deleteFactura(id);
     }
 
-    // ===== NUEVO: crear factura + detalle en una sola transacción =====
+    // ===== crear factura + detalle en una sola transacción =====
     @Transactional
     public Map<String,Object> crearConDetalle(FacturaConDetalleDTO dto) {
-        // 1) crear factura con monto 0 (el trigger ya recalcula el total)
         long facturaId = facturaRepo.insertFactura(dto.getUsuarioId(), 0.0);
 
-        // 2) insertar líneas (validando campos y convirtiendo tipos)
         if (dto.getItems() != null) {
             for (DetalleLineaDTO it : dto.getItems()) {
                 if (it.getProductoId() == null || it.getCantidad() == null || it.getPrecio() == null) {
                     throw new IllegalArgumentException("productoId, cantidad y precio son obligatorios en cada línea");
                 }
                 long productoId = it.getProductoId().longValue();
-                long cantidad   = it.getCantidad().longValue();
+                int  cantidad   = Math.toIntExact(it.getCantidad().longValue()); // repo espera int
                 double precio   = it.getPrecio().doubleValue();
 
-                // OJO: el método correcto del repo es insertLinea(...)
-                detalleRepo.insertLinea(
-                    facturaId,
-                    productoId,
-                    cantidad,
-                    precio
-                );
+                detalleRepo.insertLinea(facturaId, productoId, cantidad, precio);
             }
         }
 
-        // 3) respuesta
         Map<String,Object> resp = new HashMap<>();
         resp.put("factura_id", facturaId);
         return resp;
